Lexical Analyzer
In our project, we formed a new language for which we designed a lexical analyzer. Initially, we formed the grammar of our language and our lexical analyzer will identify the tokens of any given program written in our language. It generates error if the program does not follow the grammar. A finite automata is being designed to implement the grammar and for any code, we start from the initial state of that finite automata. If it reaches the accepting state after the end of the code we say that the tokens are valid.